// src/models/pagos/sql/proveedor.sql.ts
// Automáticamente generado a partir de molinos_create_v4.sql

export interface ProveedorSQL {
    id: number;
    nombre: string;
    id_categoria: number;
    contacto: string;
    telefono: string;
    ciudad: string;
    alias: string;
    nequi: boolean;
    compra_realizada: string;
    estado: string;
    created_at: Date;
    updated_at: Date;
}